+++ packages/client/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Node flags `isStale`, `markedForUpdate`, and `updating` are now properly set at the group level as whether any of their children have them set
 * Deprecated undocumented internal `flat` tree, will remove in next version
 * Remove unused `default` type support
+* `watchNode` now supports nested paths (e.g. `context.something`) and empty keys array
 
 # 2.43.0
 * Add `watchNode`